+++ gonk-database/src/main.rs
@@ -2,28 +2,11 @@
 use gonk_database::*;
 
 fn main() {
-    init();
+    settings::init();
+    settings::update_volume(45);
+    settings::update_queue(Vec::new(), 3232323.0);
 
-    let _song = RawSong::new(
-        "joe",
-        "joe's album",
-        "joe's title",
-        "D:\\OneDrive\\Joe\\joe's song.flac",
-        20,
-        1,
-        0.01,
-    );
-    let songs = songs_from_album(
-        "Various Artists",
-        "The Legend of Zelda: Breath of the Wild Original Soundtrack",
-    );
-    let ids: Vec<usize> = songs.iter().map(|song| song.id).collect();
-
-    bench_slow(|| {
-        let _songs = albums_by_artist("Death Grips");
-    });
-
-    bench_slow(|| {
-        let _songs = albums_by_artist("Death Grips");
-    });
+    unsafe {
+        dbg!(&settings::SETTINGS);
+    }
 }
